// src/core/domain/model/registroContacto.model.ts
export class RegistroContactoModel {
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    direccion: string;
    celular: string;
    correo: string;
    tipoDocumento: string;
    numeroDocumento: string;
    pais_emision: string;
    fechaNacimiento: Date;
    tipoContacto: 1;    
}